Extract `revision-history` to a separate spec
